fix : 일정과 user 매핑

+++ src/main/java/com/moa/moabackend/service/ScheduleService.java
@@ -4,7 +4,9 @@ import com.moa.moabackend.entity.ResponseDto;
 import com.moa.moabackend.entity.schedule.Schedule;
 import com.moa.moabackend.entity.schedule.ScheduleRequestDto;
 import com.moa.moabackend.entity.schedule.ScheduleResponseDto;
+import com.moa.moabackend.entity.user.User;
 import com.moa.moabackend.repository.ScheduleRepository;
+import com.moa.moabackend.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -12,11 +14,13 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class ScheduleService {
     private final ScheduleRepository scheduleRepository;
+    private final UserRepository userRepository;
 
     // calendar
     public ResponseDto<List<ScheduleResponseDto>> getCalendar() {
@@ -33,13 +37,15 @@ public class ScheduleService {
     }
 
     // 일정 생성
-    public ResponseDto<ScheduleResponseDto> addSchedule(ScheduleRequestDto requestDto) {
+    public ResponseDto<ScheduleResponseDto> addSchedule(ScheduleRequestDto requestDto, User user) {
         Schedule schedule = Schedule.builder()
                 .meetingDate(LocalDate.parse(requestDto.getMeetingDate()))
                 .title(requestDto.getTitle())
                 .meetingTime(LocalTime.parse(requestDto.getMeetingTime()))
                 .location(requestDto.getLocation())
                 .content(requestDto.getContent())
+                .user(user)
+//                .userNameList(requestDto.getUserNameList())
                 .build();
         scheduleRepository.save(schedule);
         return ResponseDto.success(
@@ -49,13 +55,15 @@ public class ScheduleService {
                 .meetingTime(schedule.getMeetingTime())
                 .location(schedule.getLocation())
                 .content(schedule.getContent())
+//                .userNameList(schedule.getUserNameList())
                 .build());
     }
 
     // 일정 목록 조회
-    public ResponseDto<List<ScheduleResponseDto>> getAllSchedules() {
+    public ResponseDto<List<ScheduleResponseDto>> getAllSchedules(User user) {
         List<ScheduleResponseDto> scheduleResponseLIst = new ArrayList<>();
-        List<Schedule> schedules = scheduleRepository.findAllByOrderByMeetingDate();
+//        List<Schedule> schedules = scheduleRepository.findAllByOrderByMeetingDate();
+        List<Schedule> schedules = scheduleRepository.findAllByUser(user);
         for (Schedule schedule : schedules) {
             scheduleResponseLIst.add(
                     ScheduleResponseDto.builder()
@@ -78,20 +86,27 @@ public class ScheduleService {
                         .meetingTime(schedule.getMeetingTime())
                         .location(schedule.getLocation())
                         .content(schedule.getContent())
+//                        .userNameList(schedule.getUserNameList())
                         .build()
         );
     }
 
     // 일정 수정
-    public ResponseDto<String> updateSchedule(Long scheduleId, ScheduleRequestDto requestDto) {
+    public ResponseDto<String> updateSchedule(Long scheduleId, ScheduleRequestDto requestDto, User user) {
         Schedule schedule = scheduleRepository.findById(scheduleId).get();
+        if (user.getUserId() != schedule.getUser().getUserId()) {
+            return ResponseDto.fail(404, "사용자가 일치하지 않습니다.", "Not Found");
+        }
         schedule.update(requestDto);
         return ResponseDto.success("일정이 수정되었습니다.");
     }
 
     // 일정 삭제
-    public ResponseDto<String> deleteSchedule(Long scheduleId) {
+    public ResponseDto<String> deleteSchedule(Long scheduleId, User user) {
         Schedule schedule = scheduleRepository.findById(scheduleId).get();
+        if (user.getUserId() != schedule.getUser().getUserId()) {
+            return ResponseDto.fail(404, "사용자가 일치하지 않습니다.", "Not Found");
+        }
         scheduleRepository.delete(schedule);
         return ResponseDto.success("일정이 삭제되었습니다.");
     }
